+++ Visual Studio Projects/ZWaveJS.NET/ZWaveJS.NET/Lib.cs
@@ -1,0 +1,16 @@
+using System;
+using System.IO.Ports;
+
+namespace ZWaveJS.NET
+{
+    public class Lib
+    {
+        public static string[] SerialPorts()
+        {
+            return SerialPort.GetPortNames();
+        }
+
+    }
+}
+
+